# docs/api/api-openapi.yaml
openapi: 3.0.3
info:
  title: Fiskekort-API
  version: 1.0.0
  description: API for administrasjon og kjøp av fiskekort
servers:
  - url: /api
paths:
  /fiskekort:
    get:
      summary: List alle fiskekort for innlogget bruker
      description: Returnerer alle fiskekort (aktive, fremtidige og utgåtte) for innlogget bruker
      responses:
        '200':
          description: Liste over fiskekort
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Fiskekort'
    post:
      summary: Kjøp eller opprett nytt fiskekort
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FiskekortRequest'
      responses:
        '201':
          description: Fiskekort opprettet
  /fiskekort/{id}:
    get:
      summary: Hent detaljer om et fiskekort
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Fiskekort-detaljer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Fiskekort'
...
components:
  schemas:
    Fiskekort:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        redskap:
          type: string
        gyldig_fra:
          type: string
          format: date
        gyldig_til:
          type: string
          format: date
        status:
          type: string
          enum: [aktiv, fremtidig, utgått]
        qr_kode:
          type: string
    FiskekortRequest:
      type: object
      properties:
        type:
          type: string
        redskap:
          type: string
        gyldig_fra:
          type: string
          format: date
        gyldig_til:
          type: string
          format: date